Mutation tracking journal integration, read, and write path
Patch by Blake Eggleston & Aleksey Yeshchenko
Reviewed by Blake Eggleston, Aleksey Yeshchenko, & Abe Ratnofsky
 for CASSANDRA-20304, CASSANDRA-20305, & CASSANDRA-20308, CASSANDRA-20373

+++ src/java/org/apache/cassandra/journal/Journal.java
@@ -206,8 +206,11 @@ public class Journal<K, V> implements Shutdownable
 
     public void start()
     {
+        if (state.get() == State.NORMAL)
+            return;
+
         Invariants.require(state.compareAndSet(State.UNINITIALIZED, State.INITIALIZING),
-                              "Unexpected journal state during initialization", state);
+                              "Unexpected journal state during initialization: %s", state);
         metrics.register(flusher);
 
         deleteTmpFiles();
@@ -301,7 +304,7 @@ public class Journal<K, V> implements Shutdownable
             releaser.awaitTermination(1, TimeUnit.MINUTES);
             metrics.deregister();
             Invariants.require(state.compareAndSet(State.SHUTDOWN, State.TERMINATED),
-                                  "Unexpected journal state while trying to shut down", state);
+                                  "Unexpected journal state while trying to shut down: %s", state);
         }
         catch (InterruptedException e)
         {
